PP-10711 Refactor idempotency check
- Do idempotency check in ChargesApiResource before attempting to create
a charge
- If the idempotency check finds an entry with the same key and matching
request body, return the charge response with a 200 status code rather
that a 201
- Add @JsonIgnore annotations on the getters of ChargeCreateRequest.
This means that we will not serialise nullable properties on the request
that have default values returned by the getters. This will mean that
if we add new optional properties to the ChargeCreateRequest, the
equality check between the request stored in the `idempotency` table
before the property is added will still be deemed to be equal to a newly
made request.