bau: Set appropriate logging level
The log at error level should really be at the info level. It is thrown from
the line above when the authorisation has not been successful:
```
operationResponse.throwGatewayError();
```
An example of the error message would be
```
Worldpay authorisation response (error code: 5, error: XML failed validation: Invalid payment details : Card number not recognised
```
This is a legitimate scenario.

One could argue that normal program control flow instead of exceptions should
be used for these legitimate scenarios and I would agree but refactoring this
is beyond the scope of what time allows at the moment. The effect of this
commit would be a reduction in the noise found in our sentry alerts slack
channel.