PP-10712 Handle race condition for idempotency
Handle the potential race condition when two requests to create a charge
with the same Idempotency Key are made at about the same time, resulting
in an entry in the idempotency table not existing when the check is made
prior to creating a charge. In this case, one request will commit the
database transaction first, resulting in a RollbackException being thrown
when the other request attempts to commit its transaction to insert an
entry in the `idempotency` table which violates the UNIQUE criteria.

Handle the RollbackException by responding to the request with a 409 and
"error_identifier": "IDEMPOTENCY_KEY_USED".