PP-10831 Support cancelling payment instruments for sandbox
- Deleting stored payment details tasks, such as when an agreement is cancelled,
should not throw an exception when the payment provider is Sandbox and no further
action needs to be taken.
- Add empty implementation of deleteStoredPaymentDetails in SandboxPaymentProvider
(to override the default implementation that throws an exception)
- Delete the DeleteStoredPaymentDetailsTaskHandlerIT test class, which contained a
single test to ensure an exception is thrown when the above method has not been
implemented for a particular payment provider - this is now redundant as the method
has been implemented for Worldpay, Stripe and Sandbox.