PP-10981 Add can_retry inside state for get charge response
When getting a single charge (including for frontend), include
a can_retry property inside the state object only if the following
are both true:

• the charge has an authorisation mode of AGREEMENT
• the charge’s canRetry field is explicitly set to true or false
  (not left as null)

ChargeService is so large that adding many more lines to it runs
the risk that it may cause it to achieve sentience and take over
the world. Therefore, the code for creating an
ExternalTransactionState (the Java representation of the JSON state
object) has been moved into a factory, which contains the new
logic for deciding when to include the can_retry property. This
also makes unit testing a little easier.

ChargesApiResourceIT contains some new tests to check can_retry
is serialised correctly when we want it.