PP-11244 Add new frontend gateway account response model
In order to get away from using the GatewayAccountEntity to build JSON
responses, add a new model - FrontendGatewayAccountResponse, which is
returned in responses that include the gateway account that are
consumed by the card frontend app.

This contains the subset of the fields on a gateway account that are
consumed by the frontend app.

Remove the `FrontendView` JsonView option for the GatewayAccountEntity
as it is no longer used. A subsequent PR will deal with removing the
remaining `APIView` JsonView option.