PP-11244 Prepare to reduce number of GET gateway account endpoints
We have 2 endpoints for getting a Gateway Account by the internal ID:

- `/v1/frontend/accounts/{accountId}` returns the gateway account
including the credentials and is called by selfservice and toolbox
- `/v1/api/accounts/{accountId}` returns the same response but without
credentials. It is used sometimes on toolbox, although there is no good
reason to call it over `/v1/frontend/...`

We want to remove one of these endpoints to reduce confusion. It makes
more sense to remove `/v1/frontend/...` as this suggests the endpoint
is called by the frontend app, which it isn't.

Prepare for this deletion by making the behaviour of the
`/v1/api/accounts/{accountId}` endpoint identical to
`/v1/frontend/accounts/{accountId}`.

We will then:
- update consumers to always use `/v1/api/accounts/{accountId}`
- delete `/v1/frontend/accounts/{accountId}`