PP-11152 Convert credentials map into POJO
Add a GatewayCredentials interface, which the types representing the
credentials structure for each payment gateway inherit. Add
implementations for Worldpay, Stripe and ePDQ. Also add an
implementation for Sandbox to avoid returning null/Optional when getting
the credentials for a Sandbox account, even though there are no
credentials that can be entered for Sandbox.

Add a getter to GatewayAccountCredentialsEntity to convert the
credentials JSON map to a POJO that represents the credentials structure
for the payment gateway the credentials are for.

Add a setter to GatewayAccountCredentialsEntity to allow serializing
implementations of GatewayCredentials to a JSON map to insert into the
database.

Start using the GatewayCredentials representation rather than the
credentials map when updating Worldpay credentials.