PP-11152 Use credentials object for google pay merchant ID (#4624)
Get credentials as a GatewayCredentials object for operations involving
getting/updating Google Pay merchant ID.

Adds 2 methods to the GatewayCredentials interface
- hasCredentials to indicate whether any credentials have been populated
- getGooglePayMerchantId which will only ever be present for Worldpay
credentials, but it is returned as a root field in the gateway account
object for API responses. Having this as a method on the interface means
we don't need to have PSP specific code in places we get this field.