PP-11152 Add isForRecurringPayment to GatewayRequest
We are adding in `recurring_customer_initiated` nested credentials to
the credentials JSON for Worldpay accounts, to use instead of the
credentials at the root level for payments to set up a recurring payment
agreement.

In order to determine whether to get the one-off or recurring customer
initiated credentials, we need to know whether a payment is recurring or
not. Add `isForRecurringPayment` method to GatewayRequest interface to
tell us this.

We determine whether a payment is recurring or not by whether it has an
agreement ID set. This means that for query and refund requests, where
we get the payment from ledger, we need to consume the `agreement_id`
field on returned transactions. Add a Pact test for getting a recurring
payment from ledger to test the contract for this.