PP-11152 Use credentials POJO for building gateway requests
Use the GatewayCredentials representation of the credentials when
retrieving the credentials for use when making requests to the payment
gateways.

As part of this, make the change to get the credentials for a customer
initiated recurring Worldpay payment from the
`recurring_customer_initiated` nested credentials, rather than the root
credentials. Credentials for accounts used for Worldpay recurring
payments have already been updated on all environments to include
`recurring_customer_initiated` credentials.