PP-11152 Fix exception thrown for Smartpay credentials
An exception was thrown when a Gateway Account was retrieved that has
historic Smartpay credentials, meaning pages in selfservice could not
be loaded.

Do not throw an exception when reading the gateway credentials into an
object, and instead read them as SandboxCredentials, which will be
serialised in API responses containing credentials as an empty JSON
object.