PP-11152 Improve how we retrieve legacy Worldpay credentials
Worldpay credentials for one-off payments can be stored in the database
with the "legacy" structure where the `merchant_id`, `username` and
`password` are at the root of the JSON, or new structure where the
credentials are in a nested `one_off_customer_initiated` JSON object.

To avoid conditionals in multiple places when retrieving the one-off
payment credentials, alter
WorldpayCredentials#getOneOffCustomerInitiatedCredentials to return
a WorldpayMerchantCodeCredentials populated using the legacy credentials
if the credentials are stored in the database with the legacy structure.

This also affects API responses, meaning that in selfservice we will
be able to only look at the `one_off_customer_initiated` credentials
for displaying credentials back to users, rather than having to check
whether the credentials have the new or legacy structure.