PP-11152 Fix serialising Worldpay credentials
07a376be488195a3b2b673a72043018bc70595ee intended to make it so that
Worldpay credentials stored in the database in the legacy format are
serialised in API responses with the credentials duplicated in a nested
`one_off_customer_initiated` field so that selfservice doesn't have to
check whether the credentials use the legacy structure or not, and
always consume the nested `one_off_customer_initiated` credentials.

07a376be488195a3b2b673a72043018bc70595ee didn't make this the case for
API responses due to the fact
WorldpayCredentials#getOneOffCustomerInitiatedCredentials has the
@JsonIgnore annotation so that we don't serialise the Optional in a
strange way.

Instead, populate the `one_off_customer_initiated` credentials when
deserializing the credentials from the database using the legacy
credentials if they exist and `one_off_customer_initiated` nested
credentials do not exist.

Add an integration test for this behaviour.