PP-11152 Fix serialising Worldpay credentials
07a376b intended to make it so that Worldpay credentials stored in the
database in the legacy format are serialised in API responses with the
credentials duplicated in a nested one_off_customer_initiated field so
that selfservice doesn't have to check whether the credentials use the
legacy structure or not, and always consume the nested
one_off_customer_initiated credentials.

07a376b didn't make this the case for API responses due to the fact
WorldpayCredentials#getOneOffCustomerInitiatedCredentials has the
`@JsonIgnore` annotation so that we don't serialise the Optional in a
strange way.

Remove the `@JsonIgnore` annotations on the getters that return
optionals and instead use an ObjectMapper with the Jdk8Module registered
for serialization so that we serialize Optionals sensibly.