PP-11152 Don't serialize nulls for WorldpayCredentials (#4647)
Using the Jdk8Module to serialize the Optional getters sensibly caused
it not obey the `@JsonInclude(JsonInclude.Include.NON_NULL)` annotation
for fields with getters that return Optionals.

We don't want to include null fields, particularly when serializing the
JSON for writing to the database. We also prefer this for API responses
to maintain consistency.

Instead of registering the Jdk8Module on the ObjectMapper, add
additional getters to WorldpayCredentials that don't return Optionals
for Optional fields. Add `@JsonIgnore` to the Optional getters so the
other getters are used instead.