PP-11152 Stop including legacy Worldpay creds in API response
We now always return `one_off_customer_initated` in the credentials JSON
in API responses, if the credentials are stored using the new or legacy
structure for one-off Worldpay credentials.

All consumers of the API now use the `one_off_customer_initated` from
the API response. Stop returning `merchant_id` and `username` at the
root of the credentials object if the credentials are stored with the
legacy structure in the database.