PP-11290 Stop reading legacy credentials
We have now copied the legacy credentials into a nested
`one_off_customer_initiated` object in the credentials jsonb stored in
the database. We also now only set new or updated credentials in this
nested object.

We can now remove code that falls back to reading the legacy credentials
set at the root of the jsonb.