PP-11152 Improve AuthUtilTest
Improve the tests in AuthUtilTest to make it so that for tests where we
are checking we retrieve the correct credentials for the payment type
- recurring/non-recurring and authorisation mode web/agreement for
recurring payments, the credentials has all 3 types of credentials set.

This improves the tests by making definitely sure we're selecting the
correct credential when they are all set.