bau: Remove EPDQ config
As part of PP-11272, we want to remove EPDQ-related environment variables. This means removing the EPDQ config which depends on EPDQ environment variables. This PR deletes the minimum required to get PP-11272 unblocked; there are certainly more that can be deleted but this can be done at a later time.