PP-11385 Extract expiry date from Stripe auth response
For Stripe wallet payments, we are not able to obtain the expiry date
for the card before authorising. The payment method, included in the
payment intent response includes the expiry date. Extract the expiry
date from the authorisation response so we can store it on the charge
for wallet payments.