PP-11393 Use Stripe SDK models for Transfers
Proof of concept of using the models in the Stripe SDK to deserialize
responses. This means that we don't have to re-implement the models
ourselves.

Stripe responses are more complex than we sometimes treat them, and
they have the idea of collections which might be pageable, and expandable
objects. Our models hide the complexity of these in a simplistic way
that make it difficult to extend our code if we need to deal with
collections and expandable objects.