PP-11393 Implement expandable fields for Stripe responses (#4722)
When making a request to Stripe, it is possible to specify fields to be
"expanded" in the response. If these fields are not expanded, Stripe
will return the ID for the resource the field refers to. If they are
expanded, the resource will be embedded in the response.

Deserialize fields that can be expandable to the StripeExpandableField
wrapper class. This has a getter to return the ID, and a getter that
returns an optional of the expanded object.

This is inspired by how the Stripe SDK deserializes responses:
https://stripe.dev/stripe-java/com/stripe/model/ExpandableFieldDeserializer.html

For now, just use this for getting the payment method from a payment
intent response.