PP-11563 Separate Apple and Google Pay code paths
We pretend that the Apple Pay and Google Pay requests are the same for
most of the code path, which makes the code difficult to understand.

Separate out the code paths at the point of calling methods on the
PaymentProvider interface, adding separate `authoriseApplePay` and
`authoriseGooglePay` methods. Make it clearer in the Worldpay code that
we use different XML templates for the gateway request for the different
wallet types, and use different data to build them.

For now, keep the shared code path in WalletService and
WalletAuthoriseService as this allows us to share some code in a fairly
sensible way. This can be refactored on further if there is a good way
to share this code whilst making it easier to follow.