PP-11563 Separate WalletPaymentInfo for Apple and Google
Requests to the authorise endpoints for apple and google pay contain
some common data, but some different data in the `payment_info` field.

Make this clear by adding ApplePayPaymentInfo, which exists on the
ApplePayAuthRequest and GooglePayPaymentInfo which exists on
WorldpayGooglePayAuthRequest and StripeGooglePayAuthRequest.

Make both these classes extend the abstract class WalletPaymentInfo
which contains shared fields and allows us to share some code for using
this base class.