PP-11563 Overhaul the Worldpay request builder
Add interface WorldpayOrderRequest to replace most of the existing
classes used to build Worldpay order request. This encapsulates all the
logic for building a GatewayOrder and should make it easier to follow
the code. This more closely mirrors how we build Stripe requests.

Will have subclasses per request type which will only contain the data
required to build the order for that request. This commit adds
WorldpayAuthoriseOrderRequest for building authorisation requests and
uses that instead of WorldpayOrderRequestBuilder.

Re-create the tests from WorldpayOrderRequestBuilderTest for building
the authorise request in WorldpayAuthoriseOrderRequestTest.