PP-11393 Fix status code for auth errors (#4734)
When the Stripe API returns an error, the GatewayResponse returned by
the authorise service does not have a GatewayError set, but the
BaseAuthoriseResponse has an AuthoriseStatus of Error.

We were treating this the same as a decline and returning a 400 response
code. PP-8693 made some changes so that authorisation errors return a
402 response rather than a 500 response. It makes sense that all
authorisation errors return a 402 rather than a 400, and this is what
is expected by the frontend code to authorise a wallet payment.

Remove handling for AuthoriseStatus.SUBMITTED, as this was only possible
for ePDQ which has now been retired.