PP-11601 Improve how we handle rejected authorisation for wallets
A 400 response status code returned from the connector API endpoint could mean the authorisation was rejected or it could also mean that the request to connector has an invalid payload, which we should handle by showing an error page.

This change entails returning an “error_identifier":"AUTHORISATION_REJECTED" in the error response for declined payments for the authorise wallet payment endpoints, so we can be sure the 400 means the payment has been declined.

Further change is expected in the frontend, to handle 400s with error_identifier AUTHORISATION_REJECTED by rendering the auth rejected page. For any other error_identifier, render the technical difficulties page.

The Jira ticket: https://payments-platform.atlassian.net/browse/PP-11601