PP-11513 Accept wallet flags on create gateway account request
Accept `allow_apple_pay` and `allow_google_pay` flags on the request to
create a gateway account, and use them to set these flags when creating
the gateway account.

Do not return the flags in the response, as we only return a sub-set
of properties in this response, and it's only worth adding them if
we want to consume these values. As this is only consumed by toolbox,
for which we don't have pact tests, don't bother doing this for now.