PP-12302 refactor connector part 1
The first iteration of removing our reliance on the `junitparams` framework, and paving the road to use `Junit5`.
This iteration addresses the simple cases and replaces the `@RunWith` annotation based execution of integration tests with inherited `@ClassRule` one.
Introduced `NewChargingITestBase` and `NewGatewayAccountResourceTestBase` classes. These shall be renamed without the New prefix once we refactored all of the integration tests.

This branch runs 6 more tests, the difference is due to changing some parameterised tests