PP-12349 Handle multiple balance_transactions for dispute (#5011)
It is possible for a Stripe dispute to have multiple
balance_transactions. This can happen for payments made using an
international card, where there might be an an adjustment
balance_transaction to account for currency conversion. Handle this
by summing the `net` value for all balance_transations on the dispute
lost event.

Throw an exception if the sum of the `net` amounts of the
balance_transactions is positive, as we expect a negative amount,
which represents a deduction from the connect account.