PP-12349 Calculate amounts for DISPUTE_LOST event
Handle the case where a dispute lost Stripe event has multiple
balance_transactions associated with it for calculating the net amount
and fee fields to send in the DISPUTE_LOST event.

Move calculating the total net amount and fee for a dispute lost
event to a new StripeDisputeCalculator, which is called from
StripeWebhookTaskHandler in order to calculate the total net amount
once to be used for both transferring the money in Stripe and creating
the DISPUTE_LOST event to send to ledger.