PP-12374 Improve Wallet Payments authorisation success metrics
With a previous PR[1], we have clarified the logs written by Connector when a
Google Pay or an Apple Pay payment is rejected, by indicating the charge
status more clearly, and then that the connection request had been
successful, despite the payment didn't go through.

With this change, we are now correcting the `successOrFailure` value being
passed on to metrics and alerts.

Previously we were passing success or failure depending on the connection to
the Payment Gateway being successful or not.

With this change, we are now passing `success` only in two cases:
- if the payment has gone through, i.e. only for a charge status of
  `AUTHORISATION SUCCESS`
- if 3D Secure authorisation is required, i.e. for a charge status of
  `AUTHORISATION_3DS_REQUIRED`

The latter happens only in a small proportion of Google Pay payments, and we
have agreed to count these as success(close enough and there will not be many
of them).

For all other charge status values, we will from now on be passing on the
`failure` value.

Further information in JIRA[2].

[1]
https://github.com/alphagov/pay-connector/pull/5176

[2]
https://payments-platform.atlassian.net/browse/PP-12374