PP-12844 Add and refactor create charge endpoint tests
Context: There is already a large number of integration tests for the create charge endpoint, and replicating each of these for the degatewayed endpoint would add significantly to the test running time. Some of these tests are for validation only so these are being moved into a single test class. This will minimise the impact on performance through reduced overall startup time and each test will run quicker as a unit test.

- Move validation tests for create charge endpoint into ChargesApiResourceValidationTest
- Use parameterised tests to avoid duplication for the gateway and degatewayed endpoint
- Inline fields so to make the inputs and outputs clearer
- Create accounts via API rather than in the database as per agreed team approach
- Use degatewayed urls in the setup for tests of degatewayed endpoints where possible.
- Subsequent PRs will add coverage for the degatewayed endpoint in respect of the remaining permutations of metadata, Moto, prefilled cardholder details, provider credentials, and zero amount.